requirements.txt
Django==4.0.5
uWSGI==2.0.20
django_environ
psycopg2-binary
Pillow